+++ src/lib_media/render/sdl_video.cpp
@@ -86,7 +86,7 @@ bool SDLVideo::processOneFrame(Data data) {
 
 	auto const now = m_clock->now();
 	auto const timestamp = pic->getTime() + PREROLL_DELAY; // assume timestamps start at zero
-	auto const delay = (Uint32)std::max<int64_t>(0, timestamp - now);
+	auto const delay = std::max<int64_t>(0, timestamp - now);
 	auto const delayInMs = clockToTimescale(delay, 1000);
 	SDL_Delay((Uint32)delayInMs);
